Trait rusty::diagnostics::DiagnosticAssessor

pub trait DiagnosticAssessor {
    fn assess(&self, d: Diagnostic) -> AssessedDiagnostic;

    fn assess_all(&self, d: Vec<Diagnostic>) -> Vec<AssessedDiagnostic> { ... }
}

the assessor determins the severity of a diagnostic this trait allows for different implementations for different usecases (e.g. default, compiler-settings, tests)

Required methods

fn assess(&self, d: Diagnostic) -> AssessedDiagnostic

determines the severity of the given diagnostic

Provided methods

fn assess_all(&self, d: Vec<Diagnostic>) -> Vec<AssessedDiagnostic>