Trait rusty::diagnostics::DiagnosticReporter

pub trait DiagnosticReporter {
    fn report(&self, diagnostics: &[AssessedDiagnostic], file_id: usize);

    fn register(&mut self, path: String, src: String) -> usize;
}

the DiagnosticReporter decides on the format and where to report the diagnostic to. possible implementations could print to either std-out, std-err or a file, etc.

Required methods

fn report(&self, diagnostics: &[AssessedDiagnostic], file_id: usize)

reports the given diagnostic

fn register(&mut self, path: String, src: String) -> usize

register the given path & src and returns an ID to indicate a relationship the given src (diagnostics for this src need to use this id)