Function rusty::compile_to_shared_pic_object

pub fn compile_to_shared_pic_object(
    source: String, 
    output: &str, 
    target: Option<String>
) -> Result<(), CompileError>

Compiles a given source string to a shared position independent object and saves the output.

Arguments

source - the source to be compiled

output - the location on disk to save the output

target - an optional llvm target triple If not provided, the machine's triple will be used.