rusty::index
Enum PouIndexEntry

pub enum PouIndexEntry {
    Program {
        name: String,
        instance_struct_name: String,
        instance_variable: Box<VariableIndexEntry>,
        linkage: LinkageType,
        location: SourceLocation,
        properties: FxHashMap<String, PropertyBlock>,
    },
    FunctionBlock {
        name: String,
        instance_struct_name: String,
        linkage: LinkageType,
        location: SourceLocation,
        super_class: Option<String>,
        interfaces: Vec<String>,
        properties: FxHashMap<String, PropertyBlock>,
    },
    Function {
        name: String,
        return_type: String,
        generics: Vec<GenericBinding>,
        linkage: LinkageType,
        is_variadic: bool,
        location: SourceLocation,
        is_generated: bool,
        is_const: bool,
    },
    Class {
        name: String,
        instance_struct_name: String,
        linkage: LinkageType,
        location: SourceLocation,
        super_class: Option<String>,
        interfaces: Vec<String>,
        properties: FxHashMap<String, PropertyBlock>,
    },
    Method {
        name: String,
        parent_name: String,
        property: Option<(String, PropertyKind)>,
        declaration_kind: DeclarationKind,
        return_type: String,
        instance_struct_name: String,
        linkage: LinkageType,
        location: SourceLocation,
    },
    Action {
        name: String,
        parent_name: String,
        instance_struct_name: String,
        linkage: LinkageType,
        location: SourceLocation,
    },
}

Variants

Program

Fields

name: String

instance_struct_name: String

instance_variable: Box<VariableIndexEntry>

linkage: LinkageType

location: SourceLocation

properties: FxHashMap<String, PropertyBlock>

FunctionBlock

Fields

name: String

instance_struct_name: String

linkage: LinkageType

location: SourceLocation

super_class: Option<String>

interfaces: Vec<String>

properties: FxHashMap<String, PropertyBlock>

Function

Fields

name: String

return_type: String

generics: Vec<GenericBinding>

linkage: LinkageType

is_variadic: bool

location: SourceLocation

is_generated: bool

is_const: bool

Class

Fields

name: String

instance_struct_name: String

linkage: LinkageType

location: SourceLocation

super_class: Option<String>

interfaces: Vec<String>

properties: FxHashMap<String, PropertyBlock>

Method

Fields

name: String

parent_name: String

property: Option<(String, PropertyKind)>

declaration_kind: DeclarationKind

return_type: String

instance_struct_name: String

linkage: LinkageType

location: SourceLocation

Action

Fields

name: String

parent_name: String

instance_struct_name: String

linkage: LinkageType

location: SourceLocation

Implementations

impl PouIndexEntry

pub fn get_method_name(&self) -> Option<String>

impl PouIndexEntry

pub fn get_property(&self, name: &str) -> Option<&Identifier>

pub fn get_properties(&self) -> Option<&FxHashMap<String, PropertyBlock>>

pub fn get_properties_vec(&self) -> Vec<&PropertyBlock>

pub fn create_program_entry( pou_name: &str, instance_variable: VariableIndexEntry, linkage: LinkageType, location: SourceLocation, properties: Vec<PropertyBlock>, ) -> PouIndexEntry

creates a new Program-PouIndexEntry

Arguments

name the name of the function

instance_variable the global instance-variable of the program

pub fn create_function_block_entry( pou_name: &str, linkage: LinkageType, location: SourceLocation, super_class: Option<Identifier>, interfaces: Vec<Identifier>, properties: Vec<PropertyBlock>, ) -> PouIndexEntry

creates a new FunctionBlock-PouIndexEntry

Arguments

name the name of the FunctionBlock

linkage the linkage type of the pou

pub fn create_function_entry( name: &str, return_type: &str, generic_names: &[GenericBinding], linkage: LinkageType, is_variadic: bool, location: SourceLocation, is_const: bool, ) -> PouIndexEntry

creates a new Function-PouIndexEntry

pub fn create_generated_function_entry( name: &str, return_type: &str, generic_names: &[GenericBinding], linkage: LinkageType, is_variadic: bool, location: SourceLocation, is_const: bool, ) -> PouIndexEntry

creates a new Function-PouIndexEntry generated by the compiler this will set the is_generated attribute to true.

pub fn create_action_entry( qualified_name: &str, pou_name: &str, linkage: LinkageType, location: SourceLocation, ) -> PouIndexEntry

creates a new Action-PouIndexEntry

Arguments

name the name of the action (without the pou-qualifier)

pou_name the name of the parent pou

pub fn create_class_entry( pou_name: &str, linkage: LinkageType, location: SourceLocation, super_class: Option<Identifier>, interfaces: Vec<Identifier>, properties: Vec<PropertyBlock>, ) -> PouIndexEntry

creates a new Class-PouIndexEntry

Arguments

name the name of the Class

pub fn create_method_entry( name: &str, return_type: &str, property: Option<(String, PropertyKind)>, owner_class: &str, declaration_kind: DeclarationKind, linkage: LinkageType, location: SourceLocation, ) -> PouIndexEntry

creates a new Method-PouIndexEntry

Arguments

name the name of the method (without the pou-qualifier)

return_type the name of the method’s return type

owner_class the name of the parent pou

pub fn get_name(&self) -> &str

returns the fully qualified name of this pou

pub fn get_super_class(&self) -> Option<&str>

returns the super class of this pou if supported

pub fn get_interfaces(&self) -> Vec<&str>

pub fn get_parent_pou_name(&self) -> Option<&str>

pub fn get_declaration_kind(&self) -> Option<DeclarationKind>

pub fn get_instance_struct_type_name(&self) -> Option<&str>

returns the name of the struct-type used to store the POUs state (interface-variables)

pub fn find_instance_struct_type<'idx>( &self, index: &'idx Index, ) -> Option<&'idx DataType>

returns Some(DataType) associated with this pou or None if none is associated

index the index to fetch te DataType from

pub fn get_instance_struct_type_or_void<'idx>( &self, index: &'idx Index, ) -> &'idx DataType

returns the struct-datatype associated with this pou or void if none is associated

index the index to fetch te DataType from

pub fn get_container(&self) -> &str

returns the name of the POUs container it has no container (PROGRAM, FUNCTION, etc.)

Actions and Methods return their host-POUs name

pub fn find_implementation<'idx>( &self, index: &'idx Index, ) -> Option<&'idx ImplementationIndexEntry>

returns the ImplementationIndexEntry associated with this POU

pub fn get_linkage(&self) -> &LinkageType

returns the linkage type of this pou

pub fn is_generic(&self) -> bool

returns true if this pou is a function with generic parameters, otherwise false

pub fn is_property(&self) -> bool

pub fn get_return_type(&self) -> Option<&str>

pub fn is_variadic(&self) -> bool

returns true if this pou has a variadic (last) parameter

pub fn is_action(&self) -> bool

pub fn is_program(&self) -> bool

pub fn is_function(&self) -> bool

pub fn is_void_function(&self) -> bool

pub fn is_function_block(&self) -> bool

pub fn is_class(&self) -> bool

pub fn is_method(&self) -> bool

pub fn is_stateful(&self) -> bool

pub fn is_builtin(&self) -> bool

pub fn get_location(&self) -> &SourceLocation

pub fn get_qualified_name(&self) -> Vec<&str>

Returns the POU’s without the qualifier

pub fn get_flat_reference_name(&self) -> &str

Returns the POU’s identifier without the qualifier