Struct rusty::index::Index

pub struct Index { /* fields omitted */ }

The global index of the rusty-compiler

The index contains information about all referencable elements.

Implementations

impl Index

pub fn new() -> Index

pub fn import(&mut self, other: Index)

imports all entries from the given index into the current index

imports all global_variables, member_variables, types and implementations

Arguments

other the other index. The elements are drained from the given index and moved into the current one

pub fn get_void_type(&self) -> &DataType

pub fn find_global_variable(&self, name: &str) -> Option<&VariableIndexEntry>

pub fn find_member(
&self,
pou_name: &str,
variable_name: &str
) -> Option<&VariableIndexEntry>

pub fn find_local_members(
&self,
container_name: &str
) -> Vec<&VariableIndexEntry>

pub fn is_declared_parameter(&self, pou_name: &str, index: u32) -> bool

Returns true if the current index is a VAR_INPUT, VAR_IN_OUT or VAR_OUTPUT that is not a variadic argument

pub fn find_input_parameter(
&self,
pou_name: &str,
index: u32
) -> Option<&VariableIndexEntry>

pub fn find_variable(
&self,
context: Option<&str>,
segments: &[String]
) -> Option<&VariableIndexEntry>

pub fn find_type(&self, type_name: &str) -> Option<&DataType>

pub fn get_type(&self, type_name: &str) -> Result<&DataType, CompileError>

pub fn find_effective_type<'ret>(
&'ret self,
data_type: &'ret DataTypeInformation
) -> Option<&'ret DataTypeInformation>

Retrieves the “Effective” type behind this datatype An effective type will be any end type i.e. Structs, Integers, Floats, String and Array

pub fn find_return_variable(
&self,
pou_name: &str
) -> Option<&VariableIndexEntry>

pub fn find_return_type(&self, pou_name: &str) -> Option<&DataType>

pub fn find_type_information(
&self,
type_name: &str
) -> Option<DataTypeInformation>

pub fn get_type_information(
&self,
type_name: &str
) -> Result<DataTypeInformation, CompileError>

pub fn get_types(&self) -> &IndexMap<String, DataType>

pub fn get_globals(&self) -> &IndexMap<String, VariableIndexEntry>

pub fn get_implementations(&self) -> &IndexMap<String, ImplementationIndexEntry>

pub fn register_implementation(&mut self, call_name: &str, type_name: &str)

pub fn find_implementation(
&self,
call_name: &str
) -> Option<&ImplementationIndexEntry>

pub fn register_member_variable(
&mut self,
member_info: &MemberInfo<'_>,
initial_value: Option<Statement>,
source_location: SourceRange,
location: u32
)

registers a member-variable of a container to be accessed in a qualified name. e.g. “POU.member”, “StructName.member”, etc.

#Arguments

container_name- the name of hosting container (pou or struct)

variable_name - the name of the member variable

variable_linkage - the linkage-type of that variable (one of local, global, etc. )

variable_type_name - the variable’s data type as a string

initial_value - the initial value as defined in the AST

location - the location (index) inside the container

pub fn register_global_variable(
&mut self,
name: &str,
type_name: &str,
initial_value: Option<Statement>,
source_location: SourceRange
)

pub fn register_global_variable_with_name(
&mut self,
association_name: &str,
variable_name: &str,
type_name: &str,
initial_value: Option<Statement>,
source_location: SourceRange
)

pub fn print_global_variables(&self)

pub fn register_type(
&mut self,
type_name: &str,
initial_value: Option<Statement>,
information: DataTypeInformation
)

pub fn find_callable_instance_variable(
&self,
context: Option<&str>,
reference: &[String]
) -> Option<&VariableIndexEntry>