rusty::index
Struct InterfaceIndexEntry

pub struct InterfaceIndexEntry {
    pub ident: Identifier,
    pub location: SourceLocation,
    pub methods: Vec<String>,
    pub extensions: Vec<Identifier>,
    pub properties: Vec<PropertyBlock>,
}

Fields
ident: Identifier

The interface identifier, consisting of its name and name-location

location: SourceLocation

The location of the interface as a whole

methods: Vec<String>

A list of qualified names of the methods in this interface; the actual methods are located in Index::pous

extensions: Vec<Identifier>

A list of other interfaces this interface extends

properties: Vec<PropertyBlock>

A list of the properties declared in this interface

Implementations

impl InterfaceIndexEntry

pub fn get_name(&self) -> &str

pub fn get_name_location(&self) -> &SourceLocation

pub fn get_location(&self) -> &SourceLocation

pub fn get_declared_methods<'idx>( &self, index: &'idx Index, ) -> Vec<&'idx PouIndexEntry>

Returns a list of methods this interface declared

pub fn get_derived_methods<'idx>( &'idx self, index: &'idx Index, ) -> Vec<&'idx PouIndexEntry>

Returns a list of methods this interface inherited

pub fn get_methods<'idx>( &'idx self, index: &'idx Index, ) -> Vec<&'idx PouIndexEntry>

Returns a list of methods defined in this interface, including inherited methods from derived interfaces

pub fn get_extensions(&self) -> Vec<&Identifier>

Returns a list of interfaces this interface implements

pub fn get_derived_interfaces<'idx>( &self, index: &'idx Index, ) -> Vec<Result<&'idx InterfaceIndexEntry, Identifier>>

Returns a list of interfaces this interface inherited directly

pub fn get_derived_interfaces_recursive<'i>( &self, index: &'i Index, ) -> Vec<&'i InterfaceIndexEntry>

Returns a list of ALL existing interfaces this interface inherited directly or indirectly