rusty::index

Struct TypeIndex

pub struct TypeIndex { /* private fields */ }

the TypeIndex carries all types. it is extracted into its seaprate struct so it can be internally borrowed individually from the other maps

Implementations

impl TypeIndex

pub fn find_type(&self, type_name: &str) -> Option<&DataType>

pub fn find_pou_type(&self, type_name: &str) -> Option<&DataType>

pub fn find_effective_type_by_name(&self, type_name: &str) -> Option<&DataType>

pub fn get_effective_type_by_name(&self, type_name: &str) -> &DataType

pub fn get_type(&self, type_name: &str) -> Result<&DataType, Diagnostic>

pub fn find_effective_type<'ret>( &'ret self, data_type: &'ret DataType, ) -> Option<&'ret DataType>

Retrieves the “Effective” type behind this datatype An effective type will be any end type i.e. Structs, Integers, Floats, String and Array