rusty::lowering

Module helper

Functions

create_assignment

create_assignment_paths

Takes some expression such as bar := (baz := (qux := ADR(val)), baz2 := (qux := ADR(val))) returning all final assignment paths such as [bar.baz.qux := ADR(val), bar.baz2.qux := ADR(val)].

create_assignment_with_index

create_assignments_from_initializer

Takes some expression such as foo : FooStruct := (bar := (baz := (qux := ADR(val)), baz2 := (qux := ADR(val)))); and returns assignments of form [foo.bar.baz.qux := ADR(val), foo.bar.baz2.qux := ADR(val)].

create_assignments_from_initializer_with_index

create_call_statement

create_member_reference

create_member_reference_with_location

create_ref_assignment

create_ref_assignment_with_index

extract_ref_call_argument

Checks if the given node is a REF(...) call and returns the argument if so. This is used to detect reference assignments that should use REF= syntax.

get_unit_name

Returns a sanitized unit name suitable for use as an identifier (e.g. in generated code)

insert_base_node

Inserts a new base node into the member reference chain. For example a call such as insert_base_node("b.c", a") will yield a.b.c.

new_constructor

new_implementation

new_pou

new_unit_constructor

new_variable