rusty::lowering

Module polymorphism

Lowering of method calls into indirect calls through virtual tables

This module is responsible for lowering / transforming method calls into method calls through the virtual table (for information regarding virtual tables refer to crate::lowering::vtable). In a nutshell it will transform a method call such as ptr^.foo() into __vtable_Fb#(ptr^.__vtable^).foo^(ptr^).

However, not all method calls must be lowered but rather the following cases:

1. Method calls within methods (and function block bodies)

The reason we want to lower all method calls within (other) methods is for the fact that a non overridden method may make use of an overridden method. For example consider

FUNCTION_BLOCK A
    METHOD getName: STRING
        getName := 'A';
    END_METHOD

    METHOD printName
        printf('name = %s$N', ADR(getName()));
    END_METHOD
END_FUNCTION_BLOCK

FUNCTION_BLOCK B EXTENDS A
    METHOD getName: STRING
        getName := 'B';
    END_METHOD

    METHOD persistName
        // Persist name to some file
    END_METHOD
END_FUNCTION_BLOCK

FUNCTION main
    VAR
        instanceA: A;
        instanceB: B;

        refInstanceA: POINTER TO A;
    END_VAR

    refInstanceA := ADR(instanceA);
    refInstanceA^.printName(); // Calls `A::printName` which calls `A::getName` yielding "name = A"

    refInstanceA := ADR(instanceB);
    refInstanceA^.printName(); // Calls `A::printName` which calls `B::getName` yielding "name = B"
END_FUNCTION

As described in the main function, the calls to printName must happen at runtime. Were that not the case then printName in A would resolve to A::getName at compile time, yielding an incorrect result for the second refInstanceA^.printName() call. Lowering the call to printName would result in printf('name = %s$N', ADR(__vtable_A#(THIS^.__vtable^).getName^(THIS^)).

2. Method calls through a pointer variable pointing to a class or function block

Essentially what is illustrated in 1. within the main function, consider:

FUNCTION main
    VAR
        name: STRING;
        instanceA: A;
        instanceB: B;

        refInstanceA: POINTER TO A;
    END_VAR

    refInstanceA := ADR(instanceA);
    name := refInstanceA^.getName(); // Calls `A::getName` yielding "name = A"

    refInstanceA := ADR(instanceB);
    name := refInstanceA^.getName(); // Calls `B::getName` yielding "name = B"
END_FUNCTION

While this is a simple example, and in theory compilers would be able to derive the correct method calls at compile time with some statical analysis, our compiler today is not able to do that. Specifically it does not know that the second refInstanceA variable is pointing at B and the pointer call could be simplified into a direct call to B::getName(). Instead, it relies on dynamic dispatch for a correct code execution. Again, this is done by accessing the virtual table and calling the function pointer behind it. In terms of ST code we would transform the calls into __vtable_A#(refInstanceA^.__vtable^).getName^(refInstanceA^).

One final thing to note, while the casting of the virtual tables into concrete types is not really interesting per-se, the upcasting from a child to its parent virtual table should at least be mentioned. That is, as long as the virtual table definitions are compatible, upcasting can be performed without any issues. Compatible here refers to the fact that the order of the member fields must be constant. More specifically, the methods must be defined in “ancestral hierarchical order”. To illustrate with the previous examples, assume we have

TYPE __vtable_A:
    STRUCT
        getName: __FPOINTER TO A.getName := ADR(A.getName);
        printName: __FPOINTER TO A.printName := ADR(A.printName);
    END_STRUCT
END_TYPE

TYPE __vtable_B:
    STRUCT
        getName: __FPOINTER TO B.getName := ADR(B.getName);             // Overridden
        printName: __FPOINTER TO A.printName := ADR(A.printName);       // Inherited
        persistName: __FPOINTER TO B.persistName := ADR(B.persistName); // New
    END_STRUCT

We can safely cast from B to A’s virtual table because the layout is compatible and it would result in persistName to be cut off in the cast. Were that not the case, e.g. if getName were to be swapped with persistName then the call to getName would silently result in calling persistName.

Structs

PolymorphicCallLowerer