rusty::lowering

Module property

This module is responsible for lowering any

[CompilationUnit::properties] into [CompilationUnit::units] and CompilationUnit::implementations

AstStatement::ReferenceExpr into AstStatement::CallStatement to call the GET or SET methods

The first step is triggered right after parsing the source code. For example assume some user wrote the following code

FUNCTION_BLOCK fb
    PROPERTY foo : DINT
        GET
            // ...
            foo := <expr>;
            // ...
        END_GET

        SET
            // ...
            <expr> := foo;
            // ...
        END_SET
    END_PROPERTY
END_FUNCTION_BLOCK

internally these GET and SET blocks will be lowered into methods because semantically <var> := fb.foo is equivalent to <var> := fb.get_foo() and fb.foo := <expr> is equivalent to fb.set_foo(<expr>). Hence the properties internal representation is as follows

FUNCTION_BLOCK fb
    METHOD __get_foo
        VAR
            foo : DINT; // Patched in by the lowerer
        END_VAR

        // ...
        foo := <expr>;
        // ...
        __get_foo := foo; // Patched in by the lowerer
    END_METHOD

    METHOD __set_foo
        VAR_INPUT
            foo : DINT; // Patched in by the lowerer
        END_VAR

        // ...
        <expr> := foo;
        // ...
    END_METHOD
END_FUNCTION_BLOCK

To then trigger these methods whenever a property is referenced in some statement, a second lowering stage needs to be applied. That lowering stage happens right after we have successfully annotated all AST nodes. Again, for example assume we have the following code

FUNCTION main
    VAR
        fbInstance : fb;
        localVariable : DINT;
    END_VAR

    fbInstance.foo := 5;                // We want this to be `fbInstance.__set_foo(5);`
    localVariable := fbInstance.foo;    // ... and this to be `localVariable := fbInstance.__get_foo();`
END_FUNCTION

Lowering these references is done by simply using the AstVisitorMut, iterating over all statements and identifying if any reference has a property annotation. If so, we distinguish between two cases:

An assignment where the left hand side is a property reference, in which case the whole right hand side needs to be wrapped in a function call as __set_<property name>(<right hand side>)

A reference, that is not the left hand side of an assignment, in which case the reference itself needs to be replaced with a function call as __get_<property name>()

Structs

PropertyLowerer

Functions

lower_to_pou

The actual logic for lowering properties into methods and their implementations counterpart