rusty::lowering

Module vtable

Virtual table generation of classes and function blocks.

This module is responsible for creating virtual tables for classes and function blocks, enabling dynamic dispatch for polymorphic use cases. In short, every method call needs to be invoked indirectly through a virtual table at runtime, which essentially is a collection of function pointers pointing to the method implementations of POUs. The process of creating these virtual tables can be broken down into three tasks:

1. Virtual Table POU Member Field

Every root, i.e. non-extended, class or function block will receive a __vtable: POINTER TO __VOID member field. For example a function block such as

FUNCTION_BLOCK A
    VAR
        one: DINT;
    END_VAR
END_FUNCTION_BLOCK

will internally expand to

FUNCTION_BLOCK A
    VAR
        __vtable: POINTER TO __VOID;
        one: DINT;
    END_VAR
END_FUNCTION_BLOCK

Note that we use a VOID type here because the actual type assigned to the __vtable member differs from POU to POU.

2. Virtual Table Struct Definition

As mentioned, virtual tables are essentially just a collection of function pointers reflecting a POU. Consider a function block A with methods foo and bar as well as a function block B extending A with methods foo (inherited) and bar (overridden) as well as baz (new). For these two function blocks we would generate the following virtual table structures:

TYPE __vtable_A:
    STRUCT
        foo: __FPOINTER TO A.foo := ADR(A.foo);
        bar: __FPOINTER TO A.bar := ADR(A.bar);
    END_STRUCT
END_TYPE

TYPE __vtable_B:
    STRUCT
        foo: __FPOINTER TO A.foo := ADR(A.foo); // inherited
        bar: __FPOINTER TO B.bar := ADR(B.bar); // overridden
        baz: __FPOINTER TO B.baz := ADR(B.baz); // new
    END_STRUCT
END_TYPE

3. Global Virtual Table Instances

The newly created __vtable member fields need to point to some instance of the corresponding virtual table. For that we will create one global variable for each virtual table. Deriving from the previous example we would generate the following two variables

VAR_GLOBAL
    __vtable_instance_A: __vtable_A;
    __vtable_instance_B: __vtable_B;
END_VAR

These global variables will later be assigned in the [crate::lowering::initializers] module.

Note that the actual lowering of method calls to make use of these virtual tables will happen in the crate::lowering::polymorphism module.

Structs

VirtualTableGenerator