rusty

Module resolver

Resolves (partial) expressions & statements and annotates the resulting types

Recursively visits all statements and expressions of a CompilationUnit and records all resulting types associated with the statement’s id.

Modules

const_evaluator

generics

Structs

AnnotationMapImpl

AstAnnotations

StringLiterals

TypeAnnotator

VisitorContext

Context object passed by the visitor Denotes the current context of expressions (e.g. the current pou, a defined context, etc.)

Enums

AutoDerefType

Dependency

MethodDeclarationType

ResolvingStrategy

StatementAnnotation

Traits

AnnotationMap