rusty::resolver

Module const_evaluator

Structs

UnresolvableConstant

a wrapper for an unresolvable const-expression with the reason why it could not be resolved

Functions

evaluate

evaluate_constants

Returns the resolved constants index and a Vec of qualified names of constants that could not be resolved. After constants have been evaluated, enum defaults are finalized. TODO: revisit enum-defaults-fixup as part of evaluate_constants after @ghaith’s changes to the initializer/constructor handling.