Function rusty::resolver::const_evaluator::evaluate_constants

pub fn evaluate_constants(index: Index) -> (Index, Vec<UnresolvableConstant>)

returns the resolved constants index and a Vec of qualified names of constants that could not be resolved.