rusty::resolver

Enum StatementAnnotation

pub enum StatementAnnotation {

    Value {
        resulting_type: String,
    },
    Argument {
        resulting_type: String,
        position: usize,
        depth: usize,
        pou: String,
    },
    Variable {
        resulting_type: String,
        qualified_name: String,
        constant: bool,
        argument_type: ArgumentType,
        auto_deref: Option<AutoDerefType>,
    },
    Function {
        return_type: String,
        qualified_name: String,
        generic_name: Option<String>,
        call_name: Option<String>,
    },
    FunctionPointer {
        return_type: String,
        qualified_name: String,
    },
    Type {
        type_name: String,
    },
    Program {
        qualified_name: String,
    },
    ReplacementAst {
        statement: AstNode,
    },
    Label {
        name: String,
    },
    Override {
        definitions: Vec<MethodDeclarationType>,
    },
    Super {
        name: String,
        interfaces: Vec<String>,
    },
    MethodDeclarations {
        declarations: FxHashMap<String, Vec<MethodDeclarationType>>,
    },
    Property {
        name: String,
    },
    None,

}

Variants

Value

an expression that resolves to a certain type (e.g. a + b –> INT)

Fields

resulting_type: String

Argument

An argument of a call statement

Fields

resulting_type: String

The resulting type of this argument

position: usize

The position of the parameter within its declared POU.

depth: usize

Inheritance depth from parameter declaration to calling context.

Given an inheritance chain A <- B <- C and instanceC(inA := 1, inB := 2, inC := 3):

inA := 1 will have a depth of 2 (declared in grandparent A)

inB := 2 will have a depth of 1 (declared in parent B)

inC := 3 will have a depth of 0 (declared in C itself)

pou: String

The POU name where this arguments parameter is declared, which may differ from the actual POU being called.

Given an inheritance chain A <- B <- C and instanceC(inA := 1, inB := 2, inC := 3):

inA := 1 will have a POU name of A

inB := 2 will have a POU name of B

inC := 3 will have a POU name of C

Variable

a reference that resolves to a declared variable (e.g. a –> PLC_PROGRAM.a)

Fields

resulting_type: String

the name of the variable’s type (e.g. "INT")

qualified_name: String

the fully qualified name of this variable (e.g. "MyFB.a")

constant: bool

denotes whether this variable is declared as a constant

argument_type: ArgumentType

denotes the variable type of this variable, hence whether it is an input, output, etc.

auto_deref: Option<AutoDerefType>

denotes wheter this variable has the auto-deref trait and if so what type

Function

a reference to a function

Fields

return_type: String

The defined return type of the function

qualified_name: String

The defined qualified name of the function

generic_name: Option<String>

Original name before generic resolution

call_name: Option<String>

The call name of the function iff it differs from the qualified name (generics)

FunctionPointer

A pointer to a function, mostly needed for polymorphism where function calls are handled indirectly with a virtual table.

Fields

return_type: String

The return type name of the function pointer

qualified_name: String

The name of the referenced function, e.g. MyFb.myMethod in POINTER TO MyFb.MyMethod := ADR(...)

Type

a reference to a type (e.g. INT)

Fields

type_name: String

Program

a reference to a program call or reference (e.g. PLC_PRG)

Fields

qualified_name: String

ReplacementAst

Fields

statement: AstNode

Label

a reference to a label in a POU

Fields

name: String

Override

A method override

Fields

definitions: Vec<MethodDeclarationType>

Super

Fields

name: String

interfaces: Vec<String>

MethodDeclarations

Fields

declarations: FxHashMap<String, Vec<MethodDeclarationType>>

Property

Fields

name: String

None

Implementations

impl StatementAnnotation

pub fn value(type_name: impl Into<String>) -> Self

Constructs a new StatementAnnotation::Value with the given type name

pub fn fnptr<T, U>(qualified_name: T, return_type_name: U) -> Self
where T: Into<String>, U: Into<String>,

Constructs a new StatementAnnotation::FunctionPointer with the qualified and return type name

pub fn create_override(definitions: Vec<MethodDeclarationType>) -> Self

pub fn with_generic_name(self, generic_name: &str) -> Self

pub fn is_const(&self) -> bool

pub fn is_alias(&self) -> bool

pub fn is_reference_to(&self) -> bool

pub fn is_auto_deref(&self) -> bool

pub fn data_type(type_name: &str) -> Self

pub fn is_property(&self) -> bool

pub fn qualified_name(&self) -> Option<&str>

pub fn is_fnptr(&self) -> bool