rusty::resolver
Module generics

Structs

GenericType

Functions

generic_name_resolver

Builds the correct generic name from the given information

get_smallest_possible_type

no_generic_name_resolver

This method returns the qualified name, but has the same signature as the generic resolver to be used in builtins