Struct rusty::resolver::AnnotationMapImpl

pub struct AnnotationMapImpl {
    pub new_index: Index,
    /* private fields */
}

Fields
new_index: Index
Implementations

impl AnnotationMapImpl

pub fn new() -> Self

creates a new empty AnnotationMap

pub fn import(&mut self, other: AnnotationMapImpl)

pub fn annotate(&mut self, s: &AstStatement, annotation: StatementAnnotation)

annotates the given statement (using it’s get_id()) with the given type-name

pub fn annotate_type_hint(
&mut self,
s: &AstStatement,
annotation: StatementAnnotation
)

pub fn annotate_hidden_function_call(&mut self, s: &AstStatement, f: AstStatement)

annotates the given statement s with the call-statement f so codegen can generate a hidden call f instead of generating s

pub fn add_generic_nature(&mut self, s: &AstStatement, nature: TypeNature)

Annotates the ast statement with its original generic nature