Struct rusty::resolver::TypeAnnotator

pub struct TypeAnnotator<'i> { /* private fields */ }

Implementations

impl<'i> TypeAnnotator<'i>

pub fn register_generic_pou_entries( &mut self, generic_function: &PouIndexEntry, return_type: &str, new_name: &str, generics: &FxHashMap<String, GenericType>, )

douplicates the given generic_function under the new_name (e.g. foo__INT__REAL)using the real datatypes for the generics as given in generics (e.g. { T=INT, U=REAL})

pub fn get_specific_function_annotation( &self, generics: &[GenericBinding], generic_qualified_name: &str, generic_return_type: &str, generic_map: &FxHashMap<String, GenericType>, generic_name_resolver: fn(_: &str, _: &[GenericBinding], _: &FxHashMap<String, GenericType>) -> String, ) -> (String, StatementAnnotation)

pub fn derive_generic_types( &self, generics: &[GenericBinding], generics_candidates: FxHashMap<String, Vec<String>>, ) -> FxHashMap<String, GenericType>

Derives the correct type for the generic call from the list of parameters

impl TypeAnnotator<'_>

pub fn annotate(&mut self, s: &AstNode, annotation: StatementAnnotation)

pub fn annotate_call_statement( &mut self, operator: &AstNode, parameters_stmt: Option<&AstNode>, ctx: &VisitorContext<'_>, )

impl<'i> TypeAnnotator<'i>

pub fn visit_unit( index: &Index, unit: &'i CompilationUnit, id_provider: IdProvider, ) -> (AnnotationMapImpl, FxIndexSet<Dependency>, StringLiterals)

annotates the given AST elements with the type-name resulting for the statements/expressions. Returns an AnnotationMap with the resulting types for all visited Statements. See AnnotationMap

pub fn update_expected_types( &mut self, expected_type: &DataType, statement: &AstNode, )

updates the expected types of statements on the right side of an assignment e.g. x : ARRAY [0..1] OF BYTE := [2,3];

pub fn visit_statement(&mut self, ctx: &VisitorContext<'_>, statement: &AstNode)