Trait rusty::resolver::AnnotationMap

pub trait AnnotationMap {

    // Required methods
    fn get(&self, s: &AstNode) -> Option<&StatementAnnotation>;
    fn get_hint(&self, s: &AstNode) -> Option<&StatementAnnotation>;
    fn get_hidden_function_call(&self, s: &AstNode) -> Option<&AstNode>;
    fn has_type_annotation(&self, s: &AstNode) -> bool;
    fn get_generic_nature(&self, s: &AstNode) -> Option<&TypeNature>;

    // Provided methods
    fn get_type_or_void<'i>(
        &'i self,
        s: &AstNode,
        index: &'i Index
    ) -> &'i DataType { ... }
    fn get_hint_or_void<'i>(
        &'i self,
        s: &AstNode,
        index: &'i Index
    ) -> &'i DataType { ... }
    fn get_type_hint<'i>(
        &self,
        s: &AstNode,
        index: &'i Index
    ) -> Option<&'i DataType> { ... }
    fn get_type<'i>(
        &'i self,
        s: &AstNode,
        index: &'i Index
    ) -> Option<&'i DataType> { ... }
    fn get_type_for_annotation<'a>(
        &self,
        index: &'a Index,
        annotation: &StatementAnnotation
    ) -> Option<&'a DataType> { ... }
    fn get_type_name_for_annotation<'a>(
        &'a self,
        annotation: &'a StatementAnnotation
    ) -> Option<&'a str> { ... }
    fn get_call_name(&self, s: &AstNode) -> Option<&str> { ... }
    fn get_qualified_name(&self, s: &AstNode) -> Option<&str> { ... }

}

Required Methods

fn get(&self, s: &AstNode) -> Option<&StatementAnnotation>

fn get_hint(&self, s: &AstNode) -> Option<&StatementAnnotation>

fn get_hidden_function_call(&self, s: &AstNode) -> Option<&AstNode>

fn has_type_annotation(&self, s: &AstNode) -> bool

fn get_generic_nature(&self, s: &AstNode) -> Option<&TypeNature>

Provided Methods

fn get_type_or_void<'i>(&'i self, s: &AstNode, index: &'i Index) -> &'i DataType

fn get_hint_or_void<'i>(&'i self, s: &AstNode, index: &'i Index) -> &'i DataType

fn get_type_hint<'i>( &self, s: &AstNode, index: &'i Index ) -> Option<&'i DataType>

fn get_type<'i>(&'i self, s: &AstNode, index: &'i Index) -> Option<&'i DataType>

fn get_type_for_annotation<'a>( &self, index: &'a Index, annotation: &StatementAnnotation ) -> Option<&'a DataType>

fn get_type_name_for_annotation<'a>( &'a self, annotation: &'a StatementAnnotation ) -> Option<&'a str>

fn get_call_name(&self, s: &AstNode) -> Option<&str>

returns the name of the callable that is refered by the given statemt or none if this thing may not be callable

fn get_qualified_name(&self, s: &AstNode) -> Option<&str>