Module Rusty::tcp_connection::configuration

Modules

client_mem

Structs

FileContent
File content structure

Enums

ConnectionType
Connections type

Constants

FILENAME
Filename of the configuration.

Statics

CONFIGURATION
Configuration memory structures and his default values.

Functions

check_ip_blacklist
This will check if the IP is on the blacklist or not.

default_connection
Default connection type of the server

default_max_clients
Default max clients of the server

default_port
Default port of the server

get_connection_type
Gets the connection type of the server.

get_max_clients
Gets the max clients of the server.

get_port
Gets the port of the server.

is_user_allowed
This will check if the ip client is on the blacklist or the server reaches te maximum client number.

load_config
This function will load all the configuration file data and puts into memory If the file does not exists then it will create it and put the default data.

show_info_conf
This function will output the configuration data into the screen.