Struct rusty::typesystem::DataType

pub struct DataType {
    pub name: String,
    pub initial_value: Option<ConstId>,
    pub information: DataTypeInformation,
    pub nature: TypeNature,
    pub location: SourceLocation,
}

Fields
name: Stringinitial_value: Option<ConstId>

the initial value defined on the TYPE-declration

information: DataTypeInformationnature: TypeNaturelocation: SourceLocation
Implementations

impl DataType

pub fn get_name(&self) -> &str

pub fn get_type_information(&self) -> &DataTypeInformation

pub fn clone_type_information(&self) -> DataTypeInformation

pub fn has_nature(&self, nature: TypeNature, index: &Index) -> bool

pub fn is_numerical(&self) -> bool

pub fn is_real(&self) -> bool

pub fn is_bit(&self) -> bool

pub fn is_internal(&self) -> bool

returns true if this type is an internal, auto-generated type

pub fn is_struct(&self) -> bool

pub fn is_array(&self) -> bool

returns true if this type is an array

pub fn is_enum(&self) -> bool

returns true if this type is an enum

pub fn is_vla(&self) -> bool

pub fn is_aggregate_type(&self) -> bool

returns true if this type is an array, struct or string

pub fn get_nature(&self) -> TypeNature

pub fn find_member(&self, member_name: &str) -> Option<&VariableIndexEntry>

pub fn get_members(&self) -> &[VariableIndexEntry]

pub fn find_declared_parameter_by_location( &self, location: u32 ) -> Option<&VariableIndexEntry>

pub fn find_variadic_member(&self) -> Option<&VariableIndexEntry>

pub fn find_return_variable(&self) -> Option<&VariableIndexEntry>

pub fn is_compatible_with_type(&self, other: &DataType) -> bool