rusty::typesystem

Struct Dimension

pub struct Dimension {
    pub start_offset: TypeSize,
    pub end_offset: TypeSize,
}

Fields
start_offset: TypeSizeend_offset: TypeSize
Implementations

impl Dimension

pub fn get_length(&self, index: &Index) -> Result<u32, String>

pub fn get_range(&self, index: &Index) -> Result<Range<i64>, String>

pub fn get_range_plus_one(&self, index: &Index) -> Result<Range<i64>, String>

Identical to [get_range] except for adding 1 to the end of the range. For example if the start and end values are 1 and 5 respectively, the range will be 1..6

Primarily used by Inkwell which calculates the array length as end - start which would generate an off-by-one error in the array size with [get_range] because ST ranges are inclusive.

pub fn get_range_inclusive( &self, index: &Index, ) -> Result<RangeInclusive<i64>, String>

pub fn is_undetermined(&self) -> bool