Module Rusty::udp_connection::configuration::client_mem

Structs

ClientItem
Structure of the Client.

Statics

CLIENT_LIST
This is the memory sector of the client list.

Functions

add_client
This will add the client into the memory vector.

add_udp_client
This will add the udp client into the memory vector

get_client_id
This searchs for the client IP and return the client index.

get_client_ip
Gets the client IP.

get_connected_clients
Gets the number of clients on the memory list.

get_udp_client_ip
Gets the udp client IP.

remove_client
This will remove the client if the ip is found on the list.

send_message_to_clients
This function will send the message to all the clients.

shutdown_client
This will disconnect the client from the connection with the server.