Function rusty::validation::validate_for_array_assignment

pub fn validate_for_array_assignment<T: AnnotationMap>(
    validator: &mut Validator,
    expressions: &[AstStatement],
    context: &ValidationContext<'_, T>
)

Finds and reports invalid ARRAY assignments where parentheses are missing yielding invalid ASTs. Specifically an invalid assignment such as x := (var1 := 1, var2 := 3, 4); where var2 is missing a ( will generate ExpressionList { Assignment {..}, ...} as the AST where each item after the first one would be handled as a seperate statement whereas the correct AST should have been Assignment { left: Reference {..}, right: ExpressionList {..}}. See also

https://github.com/PLC-lang/rusty/issues/707 and

array_validation_test.rs/array_initialization_validation