Crate rusty

Tour of Rust's interesting features as compared to common patterns in Go

Build tooling

Documentation

Package management

Testing

Handling failures (panic!, panic='abort', Result)

enum Result<T, E> {
    Ok(T),
    Err(E),
}

Generics

fn largest_i32(list: &[i32]) -> i32 {
    let mut largest = list[0];

    for &item in list.iter() {
        if item > largest {
            largest = item;
        }
    }

    largest
}

fn largest_char(list: &[char]) -> char {
    let mut largest = list[0];

    for &item in list.iter() {
        if item > largest {
            largest = item;
        }
    }

    largest
}

fn largest<T>(list: &[T]) -> T {
    let mut largest = list[0];

    for &item in list.iter() {
        if item > largest {
            largest = item;
        }
    }

    largest
}

Traits vs. Interfaces

Lifetimes to prevent dangling references

Trait Objects (and dyn)

Box, Rc, Ref, RefCell

MPSC channels & threads

Mutex (non re-entrant)

Send marker trait indicates that ownership of the type implementing Send can be transferred between threads

Sync marker trait indicates that it is safe for the type implementing Sync to be referenced from multiple threads

Good example from the book:

If you’ll only ever have homogeneous collections, using generics and trait bounds is preferable because the definitions will be monomorphized at compile time to use the concrete types. On the other hand, with the method using trait objects, one Struct instance can hold a Vec that contains a Box as well as a Box

Is Rust OO? It can emulate patterns, but as in Go this increases coupling which is never a good thing. See: https://doc.rust-lang.org/1.30.0/book/second-edition/ch17-03-oo-design-patterns.html

Look at patterns (esp. if-let, while-let): https://doc.rust-lang.org/1.30.0/book/second-edition/ch18-01-all-the-places-for-patterns.html

Structs

NewsArticle
an article

Screen

Tweet
a tweet

Traits

Draw

Summary
a generic method for summarization

Functions

largest
returns the largest item in the given list

longest
returns the longer of two given strings