Crate rusty

A St&ructured Text LLVM Frontent

RuSTy is an ST Compiler using LLVM

Features

Standard language support

Most of the IEC61131-3 standard for ST and general programing is supported.

Native compilation

A (currently) single ST files into object code using LLVM. A compiled object can be linked statically or dynamically linked with other programs using standard compiler linkers (ld, clang, gcc)

IR Output

An IR file can be generated from any given ST file in order to examin the generated LLVM IR code. For a usage guide refer to the User Documentation

Modules

builtins

codegen

expression_path

hardware_binding

index

lexer

linker

output

parser

resolver

Resolves (partial) expressions & statements and annotates the resulting types

typesystem

validation

Macros

expect_token

Enums

ConfigFormat

DebugLevel

ErrorFormat

OptimizationLevel

Target

Threads