Crate rusty

A Structured Text LLVM Frontent

RuSTy is an ST Compiler using LLVM

Features

Standard language support

Most of the IEC61131-3 standard for ST and general programing is supported.

Native compilation

A (currently) single ST files into object code using LLVM. A compiled object can be linked statically or dynamically linked with other programs using standard compiler linkers (ld, clang, gcc)

IR Output

An IR file can be generated from any given ST file in order to examin the generated LLVM IR code. For a usage guide refer to the User Documentation

Modules

cli

compile_error

index

Macros

expect_token

Structs

FilePath

SourceCode

The SourceCode unit is the smallest unit of compilation that can be passed to the compiler

Enums

Diagnostic

ErrNo

Traits

SourceContainer

SourceContainers offer source-code to be compiled via the load_source function. Furthermore it offers a location-String used when reporting diagnostics.

Functions

compile_module

Compiles the given source into a codegen::CodeGen using the provided context

compile_to_bitcode

Compiles the given source into a bitcode file

compile_to_ir

Compiles the given source into LLVM IR and returns it

compile_to_shared_object

Compiles a given source string to a dynamic non PIC object and saves the output.

compile_to_shared_pic_object

Compiles a given source string to a shared position independent object and saves the output.

compile_to_static_obj

Compiles a given source string to a static object and saves the output.

get_target_triple