rusty::codegen

Struct GeneratedModule

pub struct GeneratedModule<'ink> { /* private fields */ }

Implementations

impl<'ink> GeneratedModule<'ink>

pub fn try_from_bitcode( context: &'ink CodegenContext, path: &Path, ) -> Result<Self, CodegenError>

pub fn from_memory( context: &'ink CodegenContext, buffer: &MemoryBuffer, path: &Path, ) -> Result<Self, CodegenError>

pub fn try_from_ir( context: &'ink CodegenContext, path: &Path, ) -> Result<Self, CodegenError>

pub fn merge(self, other: GeneratedModule<'ink>) -> Result<Self, CodegenError>

pub fn persist( &self, output_dir: Option<&Path>, output_name: &str, format: FormatOption, target: &Target, optimization_level: OptimizationLevel, ) -> Result<PathBuf, CodegenError>

Persists the module into the disk based on output and target requirments If an object file should be generated, all optimizations will be executed on the object

pub fn get_unit_location(&self) -> &Path

pub fn persist_as_static_obj( &self, output: PathBuf, target: &Target, optimization_level: OptimizationLevel, ) -> Result<PathBuf, CodegenError>

Persists a given LLVM module to a static object and saves the output.

Arguments

codegen - The generated LLVM module to persist

output - the location on disk to save the output

target - an optional llvm target triple; if not provided, the machine’s triple will be used.

pub fn persist_to_shared_pic_object( &self, output: PathBuf, target: &Target, optimization_level: OptimizationLevel, ) -> Result<PathBuf, CodegenError>

Persists a given LLVM module to a shared postiion indepedent object and saves the output.

Arguments

codegen - The generated LLVM module to persist

output - the location on disk to save the output

target - an optional llvm target triple; if not provided, the machine’s triple will be used.

pub fn persist_to_shared_object( &self, output: PathBuf, target: &Target, optimization_level: OptimizationLevel, ) -> Result<PathBuf, CodegenError>

Persists the given LLVM module to a dynamic non PIC object and saves the output.

Arguments

codegen - The generated LLVM module to persits

output - the location on disk to save the output

target - llvm target triple

pub fn persist_to_bitcode( &self, output: PathBuf, ) -> Result<PathBuf, CodegenError>

Persists the given LLVM module into a bitcode file

Arguments

codegen - the genated LLVM module to persist

output - the location on disk to save the output

pub fn to_in_memory_bitcode(&self) -> Result<MemoryBuffer, CodegenError>

pub fn persist_to_ir(&self, output: PathBuf) -> Result<PathBuf, CodegenError>

Persits the given LLVM module into LLVM IR and saves it to the given output location

Arguments

codegen - The generated LLVM module to be persisted

output - The location to save the generated ir file

pub fn persist_to_string(&self) -> String

Persists the given module to a string

pub fn print_to_stderr(&self)

Prints the content of the module to the stderr

pub fn run<T, U>(&self, name: &str, params: &mut T) -> U

Runs the function given by name inside the compiled module. Returns the value returned by calling the function

pub fn run_no_param<U>(&self, name: &str) -> U

Runs the function given by name inside the compiled module. Returns the value returned by calling the function

pub fn add_global_function_mapping( &self, function_name: &str, local_function: usize, )

pub fn add_global_variable_mapping(&self, global_name: &str, local_var: usize)

pub fn set_name(&self, name: &str)