Module rusty::index

Modules

const_expressions

indexer

symbol

Structs

HardwareBinding

ImplementationIndexEntry

Index

The global index of the rusty-compiler

Label

A label represents a possible jump point in the source. It can be referenced by jump elements in the same unit

MemberInfo

TypeIndex

the TypeIndex carries all types. it is extracted into its seaprate struct so it can be internally borrowed individually from the other maps

VariableIndexEntry

Enums

ArgumentType

ImplementationType

PouIndexEntry

VariableType

Functions

get_init_fn_name

get_initializer_name

Returns a default initialization name for a variable or type

Type Aliases

FxIndexMap

Type alias for an IndexMap using the fx hashing algorithm, see https://github.com/rust-lang/rustc-hash

FxIndexSet

Type alias for a IndexSet using the fx hashing algorithm, see https://github.com/rust-lang/rustc-hash