Module rusty::index::const_expressions

Structs

ConstExpressions

InitData

Initializers which rely on code-execution/allocated memory addresses and are therefore not resolvable before the codegen stage

IntoStatementIter

Enums

ConstExpression

constant expressions registered here are wrapped behind this enum to indicate whether this expression was already (potentially) resolved or not, or if a resolving failed.

UnresolvableKind

Type Aliases

ConstId