Struct rusty::index::const_expressions::ConstExpressions

pub struct ConstExpressions { /* private fields */ }

Implementations

impl ConstExpressions

pub fn new() -> ConstExpressions

pub fn add_expression( &mut self, statement: AstStatement, target_type_name: String, scope: Option<String> ) -> ConstId

adds the const expression statement

statement: the const expression to add

target_type_name: the datatype this expression will be assigned to

scope: the scope this expression needs to be resolved in (e.g. a POU’s name)

pub fn find_expression( &self, id: &ConstId ) -> (Option<&AstStatement>, Option<&str>)

returns the expression associated with the given id together with an optional qualifier that represents the expressions scope (e.g. the host’s POU-name)

pub fn find_expression_target_type(&self, id: &ConstId) -> Option<&str>

pub fn find_const_expression(&self, id: &ConstId) -> Option<&ConstExpression>

similar to find_expression but it does not return the AstStatement directly. it returns a ConstExpression wrapper that indicates whether this expression was successfully resolved yet or not

pub fn clone( &self, id: &ConstId ) -> Option<(AstStatement, String, Option<String>)>

clones the expression in the ConstExpressions and returns all of its elements

pub fn mark_resolved( &mut self, id: &ConstId, new_statement: AstStatement ) -> Result<(), String>

marks the const-expression represented by the given id as resolvend and stores the the given new_statement as it’s resolved value.

pub fn mark_unresolvable( &mut self, id: &ConstId, reason: UnresolvableKind ) -> Result<(), String>

marks the const-expression represented by the given id as unresolvable with a given reason.

pub fn add_constant_expression( &mut self, expr: AstStatement, target_type: String, scope: Option<String> ) -> ConstId

adds the given constant expression to the constants arena and returns the ID to reference it

expr: the const expression to add

target_type: the datatype this expression will be assigned to

scope: the scope this expression needs to be resolved in (e.g. a POU’s name)

pub fn maybe_add_constant_expression( &mut self, expr: Option<AstStatement>, target_type_name: &str, scope: Option<String> ) -> Option<ConstId>

convinience-method to add the constant exression if there is some, otherwhise not use this only as a shortcut if you have an Option - e.g. an optional initializer. otherwhise use add_constant_expression

pub fn maybe_get_constant_statement( &self, id: &Option<ConstId> ) -> Option<&AstStatement>

convinience-method to query for an optional constant expression. if the given id is None, this method returns None use this only as a shortcut if you have an Option - e.g. an optional initializer. otherwhise use get_constant_expression

pub fn get_constant_statement(&self, id: &ConstId) -> Option<&AstStatement>

query the constants arena for an expression associated with the given id

pub fn get_resolved_constant_statement( &self, id: &ConstId ) -> Option<&AstStatement>

query the constants arena for a resolved expression associated with the given id. this operation returns None, if an unresolved/unresolvable expression was registered for the given id (for different behavior see get_constant_statement)

pub fn get_constant_int_statement_value( &self, id: &ConstId ) -> Result<i128, String>

query the constants arena for an expression that can be evaluated to an i128. returns an Err if no expression was associated, or the associated expression is a complex one (not a LiteralInteger)

pub fn import(&mut self, other: ConstExpressions)