Struct rusty::index::Index

pub struct Index { /* fields omitted */ }

The global index of the rusty-compiler

The index contains information about all referencable elements.

Implementations

impl Index

pub fn import(&mut self, other: Index)

imports all entries from the given index into the current index

imports all global_variables, member_variables, types and implementations

Arguments

other the other index. The elements are drained from the given index and moved into the current one

pub fn get_void_type(&self) -> &DataType

returns the void-type the NULL-statement has a type of void. void cannot be casted to any other

pub fn find_global_variable(&self, name: &str) -> Option<&VariableIndexEntry>

returns the VariableIndexEntry of the global variable with the given name

pub fn find_global_initializer(&self, name: &str) -> Option<&VariableIndexEntry>

returns the VariableIndexEntry of the global initializer with the given name

pub fn find_member(
&self,
container_name: &str,
variable_name: &str
) -> Option<&VariableIndexEntry>

return the VariableIndexEntry with the qualified name: container_name.variable_name

pub fn find_fully_qualified_variable(
&self,
fully_qualified_name: &str
) -> Option<&VariableIndexEntry>

return the VariableIndexEntry associated with the given fully qualified name using . as a delimiter. (e.g. “PLC_PRG.x”, or “MyClass.MyMethod.x”)

pub fn find_variable(
&self,
context: Option<&str>,
segments: &[&str]
) -> Option<&VariableIndexEntry>

pub fn find_enum_element(
&self,
enum_name: &str,
element_name: &str
) -> Option<&VariableIndexEntry>

returns the index entry of the enum-element element_name of the enum-type enum_name or None if the requested Enum-Type or -Element does not exist

pub fn get_container_members(
&self,
container_name: &str
) -> Vec<&VariableIndexEntry>

returns all member variables of the given container (e.g. FUNCTION, PROGRAM, STRUCT, etc.)

pub fn is_declared_parameter(&self, container_name: &str, index: u32) -> bool

returns true if the current index is a VAR_INPUT, VAR_IN_OUT or VAR_OUTPUT that is not a variadic argument In other words it returns whether the member variable at index of the given container is a possible parameter in call to it

pub fn find_input_parameter(
&self,
pou_name: &str,
index: u32
) -> Option<&VariableIndexEntry>

pub fn find_effective_type(&self, type_name: &str) -> Option<&DataType>

returns the effective DataType of the type with the given name if it exists

pub fn get_effective_type(
&self,
type_name: &str
) -> Result<&DataType, Diagnostic>

returns the effective DataType of the type with the given name or an Error

pub fn find_effective_type_info(
&self,
type_name: &str
) -> Option<&DataTypeInformation>

returns the effective DataTypeInformation of the type with the given name if it exists

pub fn get_effective_type_by_name(&self, type_name: &str) -> &DataType

returns the effective type of the type with the with the given name or the void-type if the given name does not exist

pub fn get_intrinsic_type_by_name(&self, type_name: &str) -> &DataType

returns the intrinsic type of the type with the given name or the void-type if the given name does not exist returns the real type behind aliases and subRanges (while effective_types will only resolve aliases)

pub fn get_type(&self, type_name: &str) -> Result<&DataType, Diagnostic>

pub fn get_type_or_panic(&self, type_name: &str) -> &DataType

expect a built-in type

pub fn find_return_variable(
&self,
pou_name: &str
) -> Option<&VariableIndexEntry>

pub fn find_return_type(&self, pou_name: &str) -> Option<&DataType>

pub fn get_type_information_or_void(
&self,
type_name: &str
) -> &DataTypeInformation

pub fn get_types(&self) -> &IndexMap<String, DataType>

Returns a list of types, should not be used to search for types, just to react on them

pub fn get_pou_types(&self) -> &IndexMap<String, DataType>

pub fn get_globals(&self) -> &IndexMap<String, VariableIndexEntry>

pub fn get_global_initializers(&self) -> &IndexMap<String, VariableIndexEntry>

pub fn get_members(
&self,
name: &str
) -> Option<&IndexMap<String, VariableIndexEntry>>

pub fn get_global_qualified_enums(
&self
) -> &IndexMap<String, VariableIndexEntry>

pub fn get_implementations(&self) -> &IndexMap<String, ImplementationIndexEntry>

pub fn register_implementation(
&mut self,
call_name: &str,
type_name: &str,
associated_class_name: Option<&String>,
impl_type: ImplementationType
)

pub fn find_implementation(
&self,
call_name: &str
) -> Option<&ImplementationIndexEntry>

pub fn register_member_variable(
&mut self,
member_info: MemberInfo<'_>,
initial_value: Option<ConstId>,
source_location: SourceRange,
location: u32
)

registers a member-variable of a container to be accessed in a qualified name. e.g. “POU.member”, “StructName.member”, etc.

#Arguments

container_name- the name of hosting container (pou or struct)

variable_name - the name of the member variable

variable_linkage - the linkage-type of that variable (one of local, global, etc. )

variable_type_name - the variable’s data type as a string

initial_value - the initial value as defined in the AST

location - the location (index) inside the container

pub fn register_member_entry(
&mut self,
container_name: &str,
entry: VariableIndexEntry
)

pub fn register_enum_element(
&mut self,
element_name: &str,
enum_type_name: &str,
initial_value: Option<ConstId>,
source_location: SourceRange
)

pub fn register_global_variable(
&mut self,
name: &str,
variable: VariableIndexEntry
)

pub fn register_global_initializer(
&mut self,
name: &str,
variable: VariableIndexEntry
)

pub fn register_type(&mut self, datatype: DataType)

pub fn register_pou_type(&mut self, datatype: DataType)

pub fn find_callable_instance_variable(
&self,
context: Option<&str>,
reference: &[&str]
) -> Option<&VariableIndexEntry>

pub fn get_mut_const_expressions(&mut self) -> &mut ConstExpressions

returns the mutable reference to all registered ConstExpressions

pub fn get_const_expressions(&self) -> &ConstExpressions

returns all registered ConstExpressions

pub fn find_intrinsic_type<'idx>(
&'idx self,
initial_type: &'idx DataTypeInformation
) -> &'idx DataTypeInformation

returns the intrinsic (built-in) type represented by the given type-information this will return the built-in type behind alias and range-types

pub fn find_instances(&self) -> InstanceIterator<'_>

Creates an iterator over all instances in the index

pub fn filter_instances(
&self,
inner_filter: fn(_: &VariableIndexEntry, _: &Index) -> bool
) -> InstanceIterator<'_>

Creates an iterator over all instances in the index The passed filter will stop the iterator from navigating deeper into variables that should not be considered To filter a variable from showing up at all, (not only its children), add a filter to the returned iterator

pub fn get_builtin_function(&self, name: &str) -> Option<&BuiltIn>

If the provided name is a builtin function, returns it from the builtin index