Struct rusty::index::Index

pub struct Index { /* private fields */ }

The global index of the rusty-compiler

The index contains information about all referencable elements.

Implementations

impl Index

pub fn import(&mut self, other: Index)

imports all entries from the given index into the current index

imports all global_variables, types and implementations

Arguments

other the other index. The elements are drained from the given index and moved into the current one

pub fn get_void_type(&self) -> &DataType

returns the void-type the NULL-statement has a type of void. void cannot be casted to any other

pub fn find_qualified_global_variable( &self, context: Option<&str>, name: &str, ) -> Option<&VariableIndexEntry>

returns the VariableIndexEntry of the global variable with the given name

pub fn find_global_variable(&self, name: &str) -> Option<&VariableIndexEntry>

returns the VariableIndexEntry of the global variable with the given name

pub fn find_global_initializer(&self, name: &str) -> Option<&VariableIndexEntry>

returns the VariableIndexEntry of the global initializer with the given name

pub fn find_local_member( &self, container_name: &str, variable_name: &str, ) -> Option<&VariableIndexEntry>

return the VariableIndexEntry with the qualified name: container_name.variable_name

pub fn find_member( &self, container_name: &str, variable_name: &str, ) -> Option<&VariableIndexEntry>

Searches for variable name in the given container, if not found, attempts to search for it in super classes

pub fn find_method( &self, container_name: &str, method_name: &str, ) -> Option<&PouIndexEntry>

Searches for method names in the given container, if not found, attempts to search for it in super class

pub fn find_fully_qualified_variable( &self, fully_qualified_name: &str, ) -> Option<&VariableIndexEntry>

return the VariableIndexEntry associated with the given fully qualified name using . as a delimiter. (e.g. “PLC_PRG.x”, or “MyClass.MyMethod.x”)

pub fn find_variable( &self, context: Option<&str>, segments: &[&str], ) -> Option<&VariableIndexEntry>

pub fn find_enum_variant( &self, name: &str, variant: &str, ) -> Option<&VariableIndexEntry>

Returns the index entry of the enum variant or None if it does not exist.

pub fn find_enum_variant_by_qualified_name( &self, qualified_name: &str, ) -> Option<&VariableIndexEntry>

Returns the index entry of the enum variant by its qualified name or None if it does not exist.

pub fn get_enum_variants_by_variable( &self, variable: &VariableIndexEntry, ) -> Vec<&VariableIndexEntry>

Returns all enum variants of the given variable.

pub fn find_enum_variant_in_pou( &self, pou: &str, variant: &str, ) -> Option<&VariableIndexEntry>

Tries to return an enum variant defined within a POU

pub fn get_enum_variants_in_pou(&self, pou: &str) -> Vec<&VariableIndexEntry>

Returns all enum variants defined in the given POU

pub fn get_container_members( &self, container_name: &str, ) -> &[VariableIndexEntry]

returns all member variables of the given container (e.g. FUNCTION, PROGRAM, STRUCT, etc.)

pub fn get_pou_members(&self, container_name: &str) -> &[VariableIndexEntry]

returns all member variables of the given POU (e.g. FUNCTION, PROGRAM, etc.)

pub fn find_pou_type(&self, pou_name: &str) -> Option<&DataType>

pub fn get_declared_parameters( &self, pou_name: &str, ) -> Vec<&VariableIndexEntry>

pub fn has_variadic_parameter(&self, pou_name: &str) -> bool

pub fn get_declared_parameter( &self, pou_name: &str, index: u32, ) -> Option<&VariableIndexEntry>

returns some if the current index is a VAR_INPUT, VAR_IN_OUT or VAR_OUTPUT that is not a variadic argument In other words it returns some if the member variable at index of the given container is a possible parameter in the call to it

pub fn get_variadic_member(&self, pou_name: &str) -> Option<&VariableIndexEntry>

pub fn find_input_parameter( &self, pou_name: &str, index: u32, ) -> Option<&VariableIndexEntry>

pub fn find_parameter( &self, pou_name: &str, index: u32, ) -> Option<&VariableIndexEntry>

pub fn find_effective_type_by_name(&self, type_name: &str) -> Option<&DataType>

returns the effective DataType of the type with the given name if it exists

pub fn find_effective_type<'ret>( &'ret self, data_type: &'ret DataType, ) -> Option<&'ret DataType>

returns the effective DataType of the given type if it exists

pub fn get_effective_type_by_name( &self, type_name: &str, ) -> Result<&DataType, Diagnostic>

returns the effective DataType of the type with the given name or an Error

pub fn find_effective_type_info( &self, type_name: &str, ) -> Option<&DataTypeInformation>

returns the effective DataTypeInformation of the type with the given name if it exists

pub fn get_effective_type_or_void_by_name(&self, type_name: &str) -> &DataType

returns the effective type of the type with the with the given name or the void-type if the given name does not exist

pub fn get_intrinsic_type_by_name(&self, type_name: &str) -> &DataType

returns the intrinsic type of the type with the given name or the void-type if the given name does not exist returns the real type behind aliases and subRanges (while effective_types will only resolve aliases)

pub fn get_type(&self, type_name: &str) -> Result<&DataType, Diagnostic>

pub fn find_type(&self, type_name: &str) -> Option<&DataType>

pub fn get_type_or_panic(&self, type_name: &str) -> &DataType

expect a built-in type This only returns types, not POUs as it is meant for builtins only

pub fn get_initial_value(&self, id: &Option<ConstId>) -> Option<&AstNode>

pub fn get_initial_value_for_type(&self, type_name: &str) -> Option<&AstNode>

Returns the initioal value registered for the given data_type. If the given dataType has no initial value AND it is an Alias or SubRange (referencing another type) this method tries to obtain the default value from the referenced type.

pub fn find_return_variable( &self, pou_name: &str, ) -> Option<&VariableIndexEntry>

pub fn find_return_type(&self, pou_name: &str) -> Option<&DataType>

pub fn get_type_information_or_void( &self, type_name: &str, ) -> &DataTypeInformation

pub fn get_types(&self) -> &SymbolMap<String, DataType>

Returns the map of types, should not be used to search for types –> see find_type

pub fn get_pou_types(&self) -> &SymbolMap<String, DataType>

Returns the map of pou_types, should not be used to search for pou_types –> see find_pou_type

pub fn get_globals(&self) -> &SymbolMap<String, VariableIndexEntry>

Returns the map of globals, should not be used to search for globals –> see find_global_variable

pub fn get_program_instances(&self) -> Vec<&VariableIndexEntry>

pub fn get_pous(&self) -> &SymbolMap<String, PouIndexEntry>

Returns the map of pous, should not be used to search for pous –> see find_pou

pub fn get_global_initializers(&self) -> &SymbolMap<String, VariableIndexEntry>

pub fn get_all_enum_variants(&self) -> Vec<&VariableIndexEntry>

pub fn get_implementations( &self, ) -> &FxIndexMap<String, ImplementationIndexEntry>

pub fn register_implementation( &mut self, call_name: &str, type_name: &str, associated_class_name: Option<&String>, impl_type: ImplementationType, generic: bool, location: SourceLocation, )

pub fn find_pou(&self, pou_name: &str) -> Option<&PouIndexEntry>

pub fn is_init_function(&self, pou_name: &str) -> bool

pub fn register_program( &mut self, name: &str, location: SourceLocation, linkage: LinkageType, )

pub fn register_pou(&mut self, entry: PouIndexEntry)

pub fn register_init_function(&mut self, name: &str)

pub fn find_implementation_by_name( &self, call_name: &str, ) -> Option<&ImplementationIndexEntry>

pub fn find_pou_implementation( &self, pou_name: &str, ) -> Option<&ImplementationIndexEntry>

pub fn register_member_variable( &mut self, member_info: MemberInfo<'_>, initial_value: Option<ConstId>, source_location: SourceLocation, location: u32, ) -> VariableIndexEntry

creates a member-variable of a container to be accessed in a qualified name. e.g. “POU.member”, “StructName.member”, etc.

#Arguments

container_name- the name of hosting container (pou or struct)

variable_name - the name of the member variable

variable_linkage - the linkage-type of that variable (one of local, global, etc. )

variable_type_name - the variable’s data type as a string

initial_value - the initial value as defined in the AST

location - the location (index) inside the container

pub fn register_enum_variant( &mut self, name: &str, variant: &str, initial_value: Option<ConstId>, source_location: SourceLocation, ) -> VariableIndexEntry

pub fn register_global_variable( &mut self, name: &str, variable: VariableIndexEntry, )

pub fn register_global_initializer( &mut self, name: &str, variable: VariableIndexEntry, )

pub fn register_type(&mut self, datatype: DataType)

pub fn register_pou_type(&mut self, datatype: DataType)

pub fn find_callable_instance_variable( &self, context: Option<&str>, reference: &[&str], ) -> Option<&VariableIndexEntry>

pub fn get_mut_const_expressions(&mut self) -> &mut ConstExpressions

returns the mutable reference to all registered ConstExpressions

pub fn get_const_expressions(&self) -> &ConstExpressions

returns all registered ConstExpressions

pub fn find_intrinsic_type<'idx>( &'idx self, initial_type: &'idx DataTypeInformation, ) -> &'idx DataTypeInformation

returns the intrinsic (built-in) type represented by the given type-information this will return the built-in type behind alias / range-types

pub fn find_elementary_pointer_type<'idx>( &'idx self, initial_type: &'idx DataTypeInformation, ) -> &'idx DataTypeInformation

pub fn find_instances(&self) -> InstanceIterator<'_>

Creates an iterator over all instances in the index

pub fn filter_instances( &self, inner_filter: fn(_: &VariableIndexEntry, _: &Index) -> bool, ) -> InstanceIterator<'_>

Creates an iterator over all instances in the index The passed filter will stop the iterator from navigating deeper into variables that should not be considered To filter a variable from showing up at all, (not only its children), add a filter to the returned iterator

pub fn get_builtin_function(&self, name: &str) -> Option<&BuiltIn>

If the provided name is a builtin function, returns it from the builtin index

pub fn get_type_layout(&self) -> &DataLayout

pub fn find_range_check_implementation_for( &self, range_type: &DataTypeInformation, ) -> Option<&ImplementationIndexEntry>

returns the implementation of the sub-range-check-function for a variable of the given dataType

pub fn add_label(&mut self, pou_name: &str, label: Label)

Adds a label definition for the POU

pub fn get_label(&self, pou_name: &str, label_name: &str) -> Option<&Label>

pub fn get_labels(&self, pou_name: &str) -> Option<&SymbolMap<String, Label>>