rusty::index
Struct VariableIndexEntry

pub struct VariableIndexEntry {
    pub initial_value: Option<ConstId>,
    pub argument_type: ArgumentType,
    pub data_type_name: String,
    pub location_in_parent: u32,
    pub source_location: SourceLocation,
    /* private fields */
}

Fields
initial_value: Option<ConstId>

an optional initial value of this variable

argument_type: ArgumentType

the type of variable

data_type_name: String

the variableโs datatype

location_in_parent: u32

the index of the member-variable in itโs container (e.g. struct). defautls to 0 (Single variables)

source_location: SourceLocation

the location in the original source-file

Implementations

impl VariableIndexEntry

pub fn new( name: &str, qualified_name: &str, data_type_name: &str, argument_type: ArgumentType, location_in_parent: u32, source_location: SourceLocation, ) -> Self

pub fn create_global( name: &str, qualified_name: &str, data_type_name: &str, source_location: SourceLocation, ) -> Self

pub fn set_linkage(self, linkage: LinkageType) -> Self

pub fn set_initial_value(self, initial_value: Option<ConstId>) -> Self

pub fn set_constant(self, is_constant: bool) -> Self

pub fn set_hardware_binding(self, binding: Option<HardwareBinding>) -> Self

pub fn set_varargs(self, varargs: Option<VarArgs>) -> Self

pub fn set_var_external(self, var_external: bool) -> Self

pub fn into_typed(&self, container: &str, new_type: &str) -> Self

Creates a new VariableIndexEntry from the current entry with a new container and type This is used to create new entries from previously generic entries

pub fn get_name(&self) -> &str

pub fn get_qualified_name(&self) -> &str

pub fn get_type_name(&self) -> &str

pub fn get_location_in_parent(&self) -> u32

pub fn is_return(&self) -> bool

pub fn is_local(&self) -> bool

pub fn is_temp(&self) -> bool

pub fn is_input(&self) -> bool

pub fn is_inout(&self) -> bool

pub fn is_constant(&self) -> bool

pub fn is_external(&self) -> bool

pub fn is_var_external(&self) -> bool

pub fn get_declaration_type(&self) -> ArgumentType

pub fn get_linkage(&self) -> LinkageType

pub fn is_in_unit(&self, unit: &str) -> bool

pub fn get_variable_type(&self) -> VariableType

pub fn has_hardware_binding(&self) -> bool

pub fn is_template(&self) -> bool

pub fn get_hardware_binding(&self) -> Option<&HardwareBinding>

pub fn is_parameter(&self) -> bool

pub fn is_in_parameter_by_ref(&self) -> bool

returns true for VAR_INPUT {ref} and VAR_IN_OUT

pub fn is_variadic(&self) -> bool

pub fn get_varargs(&self) -> Option<&VarArgs>

pub fn get_qualifier(&self) -> Option<&str>