rusty::parser
Module expressions_parser

Functions

parse_call_statement

parse_expression

parse_expression(): returns expression as Statement. if a parse error is encountered, the erroneous part of the AST will consist of an EmptyStatement and a diagnostic will be logged. That case is different from only an EmptyStatement returned, which does not denote an error condition.

parse_expression_list

parse_qualified_reference

parse_strict_literal_integer

Parses a literal integer without considering Signs or the Possibility of a Floating Point/ Exponent