Type Alias rusty::resolver::Initializers

pub type Initializers = FxIndexMap<String, InitAssignments>;

POUs and datatypes which require initialization via generated function call. The key corresponds to the scope in which the initializers were encountered. The value corresponds to the assignment data, with the key being the assigned variable name and value being the initializer AstNode.

Aliased Type

struct Initializers { /* private fields */ }