Enum rusty::typesystem::DataTypeInformation

pub enum DataTypeInformation {
    Struct {
        name: String,
        members: Vec<VariableIndexEntry>,
        source: StructSource,
    },
    Enum {
        name: String,
        referenced_type: String,
        variants: Vec<VariableIndexEntry>,
    },
    Array {
        name: String,
        inner_type_name: String,
        dimensions: Vec<Dimension>,
    },
    Pointer {
        name: String,
        inner_type_name: String,
        auto_deref: bool,
        is_reference_to: bool,
    },
    Integer {
        name: String,
        signed: bool,
        size: u32,
        semantic_size: Option<u32>,
    },
    Float {
        name: String,
        size: u32,
    },
    String {
        size: TypeSize,
        encoding: StringEncoding,
    },
    SubRange {
        name: String,
        referenced_type: String,
        sub_range: Range<AstNode>,
    },
    Alias {
        name: String,
        referenced_type: String,
    },
    Generic {
        name: String,
        generic_symbol: String,
        nature: TypeNature,
    },
    Void,
}

Variants

Struct

Fields

name: String

members: Vec<VariableIndexEntry>

source: StructSource

Enum

Fields

name: String

referenced_type: String

variants: Vec<VariableIndexEntry>

Array

Fields

name: String

inner_type_name: String

dimensions: Vec<Dimension>

Pointer

Fields

name: String

inner_type_name: String

auto_deref: bool

is_reference_to: bool

Denotes whether the variable was declared as REFERENCE TO, e.g. foo : REFERENCE TO DINT

Integer

Fields

name: String

signed: bool

size: u32

the number of bit stored in memory

semantic_size: Option<u32>

the numer of bits represented by this type (may differ from the num acutally stored)

Float

Fields

name: String

size: u32

String

Fields

size: TypeSize

encoding: StringEncoding

SubRange

Fields

name: String

referenced_type: String

sub_range: Range<AstNode>

Alias

Fields

name: String

referenced_type: String

Generic

Fields

name: String

generic_symbol: String

nature: TypeNature

Void

Implementations

impl DataTypeInformation

pub fn get_name(&self) -> &str

pub fn is_void(&self) -> bool

pub fn is_string(&self) -> bool

pub fn is_string_utf8(&self) -> bool

pub fn is_string_utf16(&self) -> bool

pub fn is_character(&self) -> bool

pub fn is_int(&self) -> bool

pub fn is_bool(&self) -> bool

pub fn is_pointer(&self) -> bool

pub fn is_unsigned_int(&self) -> bool

pub fn is_signed_int(&self) -> bool

pub fn is_float(&self) -> bool

pub fn is_struct(&self) -> bool

pub fn is_array(&self) -> bool

pub fn is_vla(&self) -> bool

pub fn is_enum(&self) -> bool

pub fn is_numerical(&self) -> bool

pub fn get_dimensions(&self) -> Option<usize>

pub fn get_vla_referenced_type(&self) -> Option<&str>

pub fn is_generic(&self, index: &Index) -> bool

pub fn is_reference_to(&self) -> bool

Returns true if the variable was declared as REFERENCE TO, e.g. foo : REFERENCE TO DINT.

pub fn is_aggregate(&self) -> bool

pub fn is_date_or_time_type(&self) -> bool

pub fn get_semantic_size(&self, index: &Index) -> u32

returns the number of bits of this type, as understood by IEC61131 (may be smaller than get_size(…))

pub fn get_size_in_bits(&self, index: &Index) -> u32

returns the number of bits used to store this type

pub fn get_size(&self, index: &Index) -> Bytes

pub fn get_string_character_width(&self, index: &Index) -> Bytes

Returns the String encoding’s alignment (character)

pub fn get_alignment(&self, index: &Index) -> Bytes

pub fn get_inner_array_type_name(&self) -> Option<&str>

pub fn get_inner_array_types<'a>( &'a self, types: &mut Vec<&'a DataTypeInformation>, index: &'a Index )

Recursively retrieves all type names for nested arrays.

This is needed because a nested array such as foo : ARRAY[1..5] OF ARRAY[5..10] OF DINT provides range information for [1..5] and [5..10] in two different types stored in the index.

pub fn get_inner_pointer_type_name(&self) -> Option<&str>

pub fn is_compatible_char_and_string(&self, other: &DataTypeInformation) -> bool

pub fn get_array_length(&self, index: &Index) -> Option<usize>

Returns the array length if DataTypeInformation is of variant DataTypeInformation::Array and None otherwise.

For example calling this function on ARRAY[1..5] OF DINT, ARRAY[1..2, 1..5] OF DINT and ARRAY[1..3] OF ARRAY[1..5] yields 5, 10 and 15 respectively.

pub fn get_enum_variants(&self) -> Option<&Vec<VariableIndexEntry>>